refactor!: promote Channel and collapse ItemScope post_process bag
Channel metadata is shared by RSS and JSON Feed adapters — keep it at
Html2rss::Channel like Article. ItemScope derives post-processor Context
config from channel so Selectors no longer dual-own a post_process_config
hash.
