+++ lib/html2rss/rss_builder.rb
@@ -57,7 +57,7 @@ module Html2rss
     end
 
     ##
-    # @param channel [Html2rss::RssBuilder::Channel] The channel information for the RSS feed.
+    # @param channel [Html2rss::Channel] The channel information for the RSS feed.
     # @param articles [Array<Html2rss::Article>] The list of articles to include in the RSS feed.
     # @param stylesheets [Array<Hash>] An optional array of stylesheet configurations.
     def initialize(channel:, articles:, stylesheets: [])
