+++ spec/lib/html2rss/rss_builder_spec.rb
@@ -10,7 +10,7 @@ RSpec.describe Html2rss::RssBuilder do
   end
 
   let(:channel) do
-    instance_double(Html2rss::RssBuilder::Channel,
+    instance_double(Html2rss::Channel,
                     title: 'Test Channel',
                     url: 'http://example.com',
                     description: 'A test channel',
